Split up sr_arithmetic.rs
